# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Rental Search API",
    "description": "This API provides access to publicly available data on the Australian rental market in 2026. The API endpoints and their usage are described in detail below."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    }
  },
  "paths": {
...
    "/user/login": {
      "post": {
        "description": "Log in to an existing user account. A request body containing the user credentials must be sent.",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "description": "The credentials of the user to log in.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "examples": [
                      "mike@gmail.com"
                    ]
                  },
                  "password": {
                    "type": "string",
                    "examples": [
                      "password"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Log in successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "examples": [
                        "ajsonwebtoken"
                      ]
                    },
                    "tokenType": {
                      "type": "string",
                      "examples": [
                        "Bearer"
                      ]
                    },
                    "expiresIn": {
                      "type": "number",
                      "examples": [
                        86400
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid log in request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Request body incomplete, both email and password are required"
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Log in failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Incorrect email or password"
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "examples": [
                    "Too many requests, please try again later."
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/user/{email}/profile": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "description": "Returns a user's profile information. This endpoint requires authorisation to use.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "The email of the profile to get",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "example": "mike@gmail.com"
                        },
                        "firstName": {
                          "type": "string",
                          "example": "Michael"
                        },
                        "lastName": {
                          "type": "string",
                          "example": "Jordan"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "example": "mike@gmail.com"
                        },
                        "firstName": {
                          "type": "string",
                          "example": "Michael"
                        },
                        "lastName": {
                          "type": "string",
                          "example": "Jordan"
                        },
                        "dob": {
                          "type": "string",
                          "format": "date",
                          "example": "1963-02-17"
                        },
                        "address": {
                          "type": "string",
                          "example": "123 Fake Street, Springfield"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "JWT token has expired"
                          ]
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "Invalid JWT token"
                          ]
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "Authorization header is malformed"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "User not found"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Authentication"
        ],
        "description": "Updates a User's Profile Information.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "The email of the profile to get",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "An object containing profile fields the user wishes to change.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "examples": [
                      "Michael"
                    ]
                  },
                  "lastName": {
                    "type": "string",
                    "examples": [
                      "Jordan"
                    ]
                  },
                  "dob": {
                    "type": "string",
                    "format": "date",
                    "examples": [
                      "1963-02-17"
                    ]
                  },
                  "address": {
                    "type": "string",
                    "examples": [
                      "123 Fake Street, Springfield"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "examples": [
                        "mike@gmail.com"
                      ]
                    },
                    "firstName": {
                      "type": "string",
                      "examples": [
                        "Michael"
                      ]
                    },
                    "lastName": {
                      "type": "string",
                      "examples": [
                        "Jordan"
                      ]
                    },
                    "dob": {
                      "type": "string",
                      "format": "date",
                      "examples": [
                        "1963-02-17"
                      ]
                    },
                    "address": {
                      "type": "string",
                      "examples": [
                        "123 Fake Street, Springfield"
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Cannot change other use profiles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Forbidden"
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "JWT token has expired"
                          ]
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "Invalid JWT token"
                          ]
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "Authorization header is malformed"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "User not found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "Request body incomplete: firstName, lastName, dob and address are required."
                          ]
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "Request body invalid: firstName, lastName and address must be strings only."
                          ]
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "Invalid input: dob must be a real date in format YYYY-MM-DD."
                          ]
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "Invalid input: dob must be a date in the past."
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}